backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.30
alembic==1.13.1
asyncpg==0.29.0          # PostgreSQL async driver
psycopg2-binary==2.9.9
redis==5.0.4
paho-mqtt==2.1.0          # Cliente MQTT para ESP32
websockets==12.0
openai==1.35.0            # OpenAI SDK
python-jose[cryptography]==3.3.0  # JWT
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.3.0
httpx==0.27.0
